Clamping and Origins
Added "Clamping" to platforms, behaving similarly to with climbables;
The player will not be "picked up" by the platform unless within the clamping range. As an effect of this, if the player is standing outside the clamping range, the platform will move without them. This is intended behaviour. Clamping should be used to prevent the player from being picked up by the final pixel of a moving platform, or to "drag" the player within a set boundary of the platform, and so should be set to near the platforms width (radius)

The Origins of a number of objects have been changed. The player object now has their Origin located at the Top Centre. The sword has been adjusted to work accordingly, and climbables have been similarly shifted.