Object parenting, Moving Semi-solids
Rewrote code to work under a more object oriented system.

Semi-solids and solid moving platforms are now children of a single parent object that acts as a general solid object.

Semi-solids are now capable of acting as mobile platforms. The semi-solid object o_semiSolid is treated as a parent object, however, so shouldnt move itself; rather, should act as a parent to all other semi-solids, moving or otherwise.

Cleaned up a lot of code.

Reverted momentum to "natural". This means, due to air resistance, you will not land on the platform after jumping from it.