Generic list menus
"list menus", such as a list of settings, can now be generated via code rather than needing their own object
this is done via the o_ui_list object, and created via create_menu. This will replace o_settings

Currently, this can only display the  names of each option, and will display them fairly unimpressively in the top right of the menu. this will be extended to correctly implementing the toggles from prior as well as other settings options.

This is not a final UI design, simply the tools needed to create it. Currently, the menu will be drawn using draw functions, but will be extended to take nine-sliced sprites also.