Text-based toggles
Added support for Text-based toggles, i.e, Settings that are shown as clickable text that cycle through a number of options.

Currently, the given option to demonstrate this is "resolution", but no such thing has actually been implemented. This exists only as a demonstration for the setting, and does not actually affect resolution.