Menu backgrounds
Menus can now be created with sprites, as opposed to the default semi-transparent look. These sprites should ideally be properly 9-sliced, though not mandatorily. The sprite will cycle through its animation frames, so animations are possible.
If sprite is set to -1, the original default grey rect is used.
While any shape can be used, the menu assumes a rectangular shape and so UI elements are still rendered on a rectangular surface.