Inventory UI
Added a UI for the inventory, Accessible by pressing E
This displays the inventory in a customisably placed and sized n x m grid
Added an item, "Artefact1", which is used to demonstrate the UI
The UI will show which slot is being hovered over, and display a tooltip for the item. As a testing feature, empty slots also show a tooltip, though this will be turned off in the next update.

The collectible item object now displays the sprite of the item it holds.