Doors
Added doors that require interaction to pass through, which then act as a portal

Fixed Interaction code, to now work after changing rooms too

Added a new object, o_interaction, as a parent object for interactable objects. An additional sign has been added to the testing room to demonstrate this - it is no different to the others in function.