Savefile
Added a Save system to the game. Currently, this saves the Room data to a text file, which can be loaded from the pause menu.
Loading does NOT affect the current room. It should eventually only be possible in rooms where there are no objects that need to be saved and loaded

It also currently saves all room data and only room data. It does not save player data yet.

The save room is currently hidden and one way. It cannot be exited. Once saving is properly implemented, this will be more accessible.