Fixes and platform toggles
Fixed a bug where the current room's information would not be updated when loading a save.

Added Toggle functionality to (horizontal) platforms. The current implementation is slightly backwards and should be changed when generalised, but this works for now.

Fixed platform behaviour to work as expected when a platform is deactivated