Struct-driven Controls
Rewrote all events that relied on keyboard inputs to be remappable. This involved removing reliance on keyboard events and moving these to the Step event

All controls now refer to a new struct, global.keys, that stores what key performs each action. This means all controls can be changed with a single line, and opens the door for remapping and also controller support.