Controller Support
A huge overhaul to the entire control system, adding support for Controllers. This is tested with DualShock 4 controllers, but *should* be able to work with any controller.

This includes some long overdue fixes, such as making the pause and map keys mappable and allowing keyboard controls for attacking

Known bug: Using analogue sticks in the menu does not work.