Hitbox driven attacks
Player attacks now deal damage via hitboxes.
Hitboxes can be set to either squares or circles, and have variable size. They can either follow the players movement, be stationary, or move at a constant rate.
Their lifespan can be set.

Importantly, a delay on the hitboxes activation can also be set. This allow for attacks to have multiple hitboxes in sequence, such as to align with an animation.
A basic "animation" has been added to the sword object to demonstrate this. However, this will soon be retired in favour for a more complex attacking system.