Sound, threshold and pogos
Added a sound effect to hitfreeze attacks.

Added a Threshold to analogue stick inputs, such that above a given range, inputs are considered to be their maximum.
Added the corresponding settings item to modify this.

Added Pogo-boxes, i.e, hitboxes that apply a "pogo" effect. These set the velocity of the player after successfully landing an attack, for example to "bounce" off an enemy